Refactor ES ITCase.
