+++ bitsail-connectors/connector-elasticsearch/pom.xml
@@ -49,20 +49,5 @@
             <version>${elasticsearch.version}</version>
             <scope>compile</scope>
         </dependency>
-
-        <!-- test -->
-        <dependency>
-            <groupId>com.bytedance.bitsail</groupId>
-            <artifactId>bitsail-connector-test</artifactId>
-            <version>${revision}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.bytedance.bitsail</groupId>
-            <artifactId>bitsail-connector-fake</artifactId>
-            <version>${revision}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
-
 </project>
