+++ bitsail-connectors/connector-elasticsearch/src/test/resources/es_doc_parameter_test.json
@@ -13,42 +13,34 @@
       "id_delimiter": ",",
       "columns": [
         {
-          "index": 0,
           "name": "id",
           "type": "bigint"
         },
         {
-          "index": 1,
           "name": "varchar_type",
           "type": "string"
         },
         {
-          "index": 2,
           "name": "text_type",
           "type": "string"
         },
         {
-          "index": 3,
           "name": "bigint_type",
           "type": "bigint"
         },
         {
-          "index": 4,
           "name": "date",
           "type": "string"
         },
         {
-          "index": 5,
           "name": "index",
           "type": "string"
         },
         {
-          "index": 6,
           "name": "op_type",
           "type": "string"
         },
         {
-          "index": 7,
           "name": "version",
           "type": "long"
         }
