+++ bitsail-connectors/connector-elasticsearch/src/test/resources/es_writer_parameter_test.json
@@ -4,27 +4,22 @@
       "es_id_fields": "id",
       "columns": [
         {
-          "index": 0,
           "name": "id",
           "type": "string"
         },
         {
-          "index": 1,
           "name": "varchar_type",
           "type": "string"
         },
         {
-          "index": 2,
           "name": "text_type",
           "type": "string"
         },
         {
-          "index": 3,
           "name": "bigint_type",
           "type": "bigint"
         },
         {
-          "index": 4,
           "name": "date",
           "type": "string"
         }
