+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-cdc/pom.xml
@@ -62,12 +62,14 @@
             <groupId>org.testcontainers</groupId>
             <artifactId>kafka</artifactId>
             <version>1.17.2</version>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
             <groupId>org.apache.kafka</groupId>
             <artifactId>kafka-clients</artifactId>
             <version>${kafka.clients.version}</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
