+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-elasticsearch/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>bitsail-test-integration-kafka</artifactId>
+    <artifactId>bitsail-test-integration-elasticsearch</artifactId>
 
     <dependencies>
         <dependency>
@@ -37,23 +37,30 @@
 
         <dependency>
             <groupId>com.bytedance.bitsail</groupId>
-            <artifactId>connector-kafka</artifactId>
+            <artifactId>connector-elasticsearch</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
 
-        <!-- deps for create kafka container -->
+        <!-- elasticsearch data source -->
         <dependency>
             <groupId>org.testcontainers</groupId>
-            <artifactId>kafka</artifactId>
-            <version>1.17.2</version>
+            <artifactId>elasticsearch</artifactId>
+            <version>1.17.3</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>org.apache.kafka</groupId>
-            <artifactId>kafka-clients</artifactId>
-            <version>${kafka.clients.version}</version>
+            <groupId>org.elasticsearch.client</groupId>
+            <artifactId>elasticsearch-rest-client</artifactId>
+            <version>${elasticsearch.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.elasticsearch.client</groupId>
+            <artifactId>elasticsearch-rest-high-level-client</artifactId>
+            <version>${elasticsearch.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
