+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-elasticsearch/src/test/resources/es_sink_test.json
@@ -4,7 +4,7 @@
       "cid": 0,
       "domain": "test",
       "job_id": -25,
-      "job_name": "bitsail_connector_es_batch_sink_test",
+      "job_name": "bitsail_test_integration_fake_to_elasticsearch",
       "instance_id": -205,
       "user_name": "root",
       "checkpoint_interval": "10000",
@@ -18,22 +18,18 @@
       "unique_fields": "id",
       "columns": [
         {
-          "index": 0,
           "name": "id",
           "type": "string"
         },
         {
-          "index": 1,
           "name": "varchar_type",
           "type": "string"
         },
         {
-          "index": 2,
           "name": "text_type",
           "type": "string"
         },
         {
-          "index": 3,
           "name": "bigint_type",
           "type": "bigint"
         }
@@ -44,22 +40,18 @@
       "es_id_fields": "id",
       "columns": [
         {
-          "index": 0,
           "name": "id",
           "type": "string"
         },
         {
-          "index": 1,
           "name": "varchar_type",
           "type": "string"
         },
         {
-          "index": 2,
           "name": "text_type",
           "type": "string"
         },
         {
-          "index": 3,
           "name": "bigint_type",
           "type": "bigint"
         }
